+++ rollup.config.mjs
@@ -22,7 +22,7 @@ export default [
     ],
     output: {
       format: 'umd',
-      name: 'readExcelFile',
+      name: 'readXlsxFile',
       file: 'bundle/read-excel-file.min.js',
       sourcemap: true,
       globals: {